db4o.net update
git-svn-id: http://source.db4o.com/db4o/trunk/db4o.net@14996 915b37e7-0715-0410-8bc3-b14b59444072

+++ Db4objects.Db4o.Tests/Db4objects.Db4o.Tests/Common/Handlers/LongHandlerTestCase.cs
@@ -20,7 +20,7 @@ namespace Db4objects.Db4o.Tests.Common.Handlers
 		public virtual void TestReadWrite()
 		{
 			MockWriteContext writeContext = new MockWriteContext(Db());
-			long expected = unchecked((int)(0x1020304050607080l));
+			long expected = unchecked((long)(0x1020304050607080l));
 			LongHandler().Write(writeContext, expected);
 			MockReadContext readContext = new MockReadContext(writeContext);
 			long longValue = (long)LongHandler().Read(readContext);
@@ -29,8 +29,8 @@ namespace Db4objects.Db4o.Tests.Common.Handlers
 
 		public virtual void TestStoreObject()
 		{
-			LongHandlerTestCase.Item storedItem = new LongHandlerTestCase.Item(unchecked((int
-				)(0x1020304050607080l)), unchecked((int)(0x1122334455667788l)));
+			LongHandlerTestCase.Item storedItem = new LongHandlerTestCase.Item(unchecked((long
+				)(0x1020304050607080l)), unchecked((long)(0x1122334455667788l)));
 			DoTestStoreObject(storedItem);
 		}
